refactor: service 패키지 정리, Todo 리스트 임시작성
