+++ src/main/java/com/devnity/devnity/domain/mapgakco/controller/MapgakcoCommnetV1Controller.java
@@ -1,6 +1,6 @@
 package com.devnity.devnity.domain.mapgakco.controller;
 
-import com.devnity.devnity.domain.mapgakco.service.MapgakcoCommentService;
+import com.devnity.devnity.domain.mapgakco.service.mapgakcocomment.MapgakcoCommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
