+++ src/main/java/com/devnity/devnity/domain/mapgakco/controller/MapgakcoV1Controller.java
@@ -4,7 +4,7 @@ import com.devnity.devnity.common.api.ApiResponse;
 import com.devnity.devnity.domain.auth.jwt.JwtAuthentication;
 import com.devnity.devnity.domain.mapgakco.dto.mapgakco.request.MapgakcoCreateRequest;
 import com.devnity.devnity.domain.mapgakco.entity.MapgakcoStatus;
-import com.devnity.devnity.domain.mapgakco.service.MapgakcoService;
+import com.devnity.devnity.domain.mapgakco.service.mapgakco.MapgakcoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
